Tag Controller tests moved to folder Test
