+++ tests/Feature/Controller/Tag/TagListControllerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Feature\Controller;
+namespace Tests\Feature\Controller\Tag;
 
 use Tests\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
